changed the suffix of all tests that extend AbstractRedisInstrumentationTest from Test to IT.
