+++ apm-agent-plugins/apm-redis-plugin/apm-jedis-plugin/src/test/java/co/elastic/apm/agent/jedis/Jedis1InstrumentationIT.java
@@ -18,47 +18,41 @@
  */
 package co.elastic.apm.agent.jedis;
 
+import co.elastic.apm.agent.redis.AbstractRedisInstrumentationTest;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import redis.clients.jedis.BinaryJedis;
-import redis.clients.jedis.JedisShardInfo;
-import redis.clients.jedis.ShardedJedis;
-
-import java.util.List;
+import redis.clients.jedis.Jedis;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class Jedis2InstrumentationTest extends Jedis1InstrumentationTest {
+class Jedis1InstrumentationIT extends AbstractRedisInstrumentationTest {
 
-    private ShardedJedis shardedJedis;
-    private BinaryJedis binaryJedis;
+    protected Jedis jedis;
 
     @BeforeEach
-    void setUp() {
-        shardedJedis = new ShardedJedis(List.of(new JedisShardInfo("localhost", redisPort)));
-        binaryJedis = new BinaryJedis("localhost", redisPort);
+    void setUpJedis() {
+        jedis = new Jedis("localhost", redisPort);
     }
 
     @AfterEach
-    void tearDown() {
-        shardedJedis.close();
-        binaryJedis.close();
+    void tearDownJedis() {
+        try {
+            // this method does not exist in Jedis 1
+            Jedis.class.getMethod("close").invoke(jedis);
+        } catch (NoSuchMethodException e) {
+            // ignore, this version of redis does not support close
+        } catch (ReflectiveOperationException e) {
+            e.printStackTrace();
+        }
     }
 
     @Test
-    void testShardedJedis() {
-        shardedJedis.set("foo", "bar");
-        assertThat(shardedJedis.get("foo".getBytes())).isEqualTo("bar".getBytes());
+    void testJedis() {
+        jedis.set("foo", "bar");
+        assertThat(jedis.get("foo".getBytes())).isEqualTo("bar".getBytes());
 
         assertTransactionWithRedisSpans("SET", "GET");
     }
 
-    @Test
-    void testBinaryJedis() {
-        binaryJedis.set("foo".getBytes(), "bar".getBytes());
-        assertThat(binaryJedis.get("foo".getBytes())).isEqualTo("bar".getBytes());
-
-        assertTransactionWithRedisSpans("SET", "GET");
-    }
 }
